Render latest snippets on home page
This commit updates the `home` handler to fetch the latest snippets and
render them using templates.

It also introduces a new `templateData` struct that can hold a slice of
snippets for the home page template. The `showSnippet` template has been
updated to use a `with` block for cleaner access to snippet data.