Add template functions and current year
Introduces a new `templateData` struct field for the current year and a
`humanDate` template helper.

The `addDefaultData` helper function is added to automatically populate
the `CurrentYear` field. This function is now called within the `render`
method to ensure the current year is always available in templates.

The `humanDate` function is registered in the template cache, allowing
dates to be formatted more readably in the UI. This is applied to the
`Created` and `Expires` fields in the `show.page.tmpl` and the `Created`
field in `home.page.tmpl`.